added witnessListButton.js and refactored code to move WitnessListButton and the event listener to that module.
