+++ scripts/criminals/CriminalList.js
@@ -55,8 +55,8 @@ export const criminalListButton = () => {
     `
 }
 /*
-*   Listens for a "click" event and dispatches the custom event, witnessListButtonClicked, to the eventHub
-*   to set the article element (.listContainer) to empty and render a list of witnesses to the DOM in (.listContainer).
+*   Listens for a "click" event and dispatches the custom event, criminalListGenerate, to the eventHub
+*   to set the article element (.listContainer) to empty and render a list of criminals to the DOM in (.listContainer).
 */
 eventHub.addEventListener(
     "click",
