+++ scripts/main.js
@@ -3,14 +3,15 @@ import { criminalList, criminalListButton } from "./criminals/CriminalList.js";
 import { ConvictionSelect } from "./convictions/ConvictionsSelect.js";
 import { getConvictions } from "./convictions/convictionsDataProvider.js";
 import { NoteForm } from "./notes/NoteForm.js";
-import { witnessListButton } from "./witnesses/WitnessList.js";
+import { witnessListButton } from "./witnesses/WitnessListButton.js";
 import { getWitnesses } from "./witnesses/witnessDataProvider.js";
 import { getNotes } from "./notes/noteDataProvider.js";
 import { noteListButton } from "./notes/NoteList.js";
+import { witnessListEventListener } from "./witnesses/WitnessList.js";
 
+getWitnesses().then(witnessListButton).then(witnessListEventListener);
 getNotes();
 NoteForm();
 getConvictions().then(ConvictionSelect);
-getWitnesses().then(witnessListButton);
 getCriminals().then(criminalList).then(criminalListButton);
 noteListButton();
